+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When printing the connected peers, print the underlay IP instead of the overlay IP.
 - The link cost of a peer is now the smoothed average. This makes sure a single short latency spike doesn't disrupt routing.
 - On Linux, set the TUN ip as /7 and avoid setting a /64 route. This brings it in line with OSX.
+- When selecting the best route for a subnet, consider the currently
+  installed route and only switch if it is significantly better, or
+  directly connected
 
 ### Fixed